Add test_localization.py
